requirements.txt
streamlit>=1.32
python-dotenv>=1.0
pyyaml>=6.0
